Fixed bug if no functions have been previously made in fish

+++ setup.py
@@ -3,6 +3,8 @@ import os
 
 def getShell():
     shell = input("Are you using a fish shell? (y, n) >> ")
+    global dl_path
+    dl_path = os.path.dirname(os.path.realpath("./")) + '/daylog/daylog/daylog.py'
     if shell == "y":
         fishInstall()
     else:
@@ -10,24 +12,28 @@ def getShell():
 
 def fishInstall():
     homeRoot = os.path.expanduser('~')
-    dl_path = os.path.dirname(os.path.realpath("./")) + '/daylog/daylog/daylog.py'
-    fish_path = homeRoot + "/.config/fish/functions/daylog.fish"
-    def installOnFish():
-        fish_write = '''function daylog
+    fish_check = homeRoot + "/.config/fish/functions"
+    if os.path.exists(fish_check) is False:
+        os.makedirs(fish_check, exist_ok=True)
+    else:
+        fish_path = homeRoot + "/.config/fish/functions/daylog.fish"
+        if os.path.getsize(fish_path) == 0: 
+            installOnFish(fish_path)
+        else:
+            os.remove(fish_path)
+            installOnFish(fish_path)
+
+def installOnFish(fish_path):
+    fish_write = '''function daylog
 {dir} $argv
 end'''.format(dir = dl_path)
-        with open(fish_path, "a") as f:
-            try:
-                f.write(fish_write)
-                print("daylog installed on fish")
-            except: 
-                print("Error in writing to " + fish_path)
+    with open(fish_path, "a") as f:
+        try:
+            f.write(fish_write)
+            print("daylog installed on fish")
+        except: 
+            print("Error in writing to " + fish_path)
 
-    if os.path.getsize(fish_path) == 0: 
-       installOnFish()
-    else:
-        os.remove(fish_path)
-        installOnFish()
 
 def bashInstall():
     homeRoot = os.path.expanduser('~')
